% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeTS.R
\name{analyzeTS}
\alias{analyzeTS}
\title{analyzeTS}
\usage{
analyzeTS(x, y)
}
\arguments{
\item{x}{Object of class \emph{matrix} or \emph{data.frame}.}

\item{y}{Vector of class \emph{character} or \emph{numeric} with a length equal to the number of rows in \emph{x}.}
}
\value{
A \emph{SpatialPointsDataDrame} with the coordinate pairs for each of the sampeled pixels.
}
\description{
Extracts time series data from a \emph{RasterStack} for a \emph{SpatialPoints} object.
}
\details{
{For each unique value in \emph{y}, the function will select the rows in \emph{x} that correspond to it and estimate the
median, Median Absolte Deviation (MAD), minimum, maximum, mean and standard deviation for each column. Then, the function will build
a plot showing the median and draw a buffer that expresses the minimum and maximum. Furthermore, the function will cross-correlate each
row in \emph{x} against the median values for each unique class. The final output is a list consisting of:
\itemize{
 \item{\emph{y.statistics} - Median, minimum and maximum values for each column in \emph{x} over each unique class in \emph{y}.}
 \item{\emph{y.r2} - \eqn{R^{2}} between the each row in \emph{x} and the median values for each unique class found in \emph{y.statistics}.}
 \item{\emph{plots} - List of line plots for each unique element in \emph{y}.}}}
}
\examples{
{

require(raster)

# read raster data
r <- brick(system.file("extdata", "ndvi.tif", package="fieldRS"))

# read field data
p <- shapefile(system.file("extdata", "fields.shp", package="fieldRS"))

# derive time series
ev <- extractTS(r, p)


}
}
\seealso{
\code{\link{extractTS}} \code{\link{assignClass}} \code{\link{classModel}}
}
