% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignClass.R
\name{assignClass}
\alias{assignClass}
\title{assignClass}
\usage{
assignClass(x, y)
}
\arguments{
\item{x}{A numeric vector, a \emph{matrix} or a \emph{data.frame}.}

\item{y}{A numeric vector, a \emph{matrix} or a \emph{data.frame}.}
}
\value{
A list containing containing multi-variate information on the classification of \emph{x}.
}
\description{
Classifies temporal profiles based on reference data.
}
\details{
{The function correlates each row in \emph{x} against each row in \emph{y} and classifies
\emph{x} based on the highest r2 value. The final output is provided as a list containing:
\itemize{
 \item{\emph{class} - Class ID of \emph{x} corresponding to the rrow in \emph{y} with the highest r2 when compared with \emph{x}.}
 \item{\emph{r2} - R2 value for the best fit.}
 \item{\emph{r2diff} - Difference between the 1st and 2nd highest correlation values.}
 \item{\emph{rmse} - Root Mean Square Error (RMSE) betwen \emph{x} and the best fitting profile in \emph{y}.}}}
}
\examples{
{

require(raster)

# read raster data
r <- brick(system.file("extdata", "ndvi.tif", package="fieldRS"))

# read field data
data(fieldData)

# read reference profiles
data(referenceProfiles)

# derive time series
ev <- extractTS(fieldData[1,], r)

# compare derived and reference profiles
assignClass(ev$weighted.mean, referenceProfiles[,2:6])

}
}
\seealso{
\code{\link{extractTS}} \code{\link{analyzeTS}}
}
