% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classFilter.R
\name{classFilter}
\alias{classFilter}
\title{classFilter}
\usage{
classFilter(x, y)
}
\arguments{
\item{x}{An object of class \emph{RasterLayer}.}

\item{y}{A single or a two element vector of class \emph{numeric}.}
}
\value{
A \emph{RasterLayer}.
}
\description{
Smoothing of a categorical raster object.
}
\details{
{For all non-NA pixels in \emph{x}, the function uses a moving window with a number of pixels defined by \emph{y} to 
look at its surroundings. If at least half of the values are non-NA, the function will return the dominant class.}
}
\examples{
{}
}
\seealso{
\code{\link{classModel}} \code{\link{ccLabel}}
}
