% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexTrace.R
\name{complexTrace}
\alias{complexTrace}
\title{complexTrace}
\usage{
complexTrace(x)
}
\arguments{
\item{x}{Object of class \emph{RasterLayer}.}
}
\value{
A \emph{SpatialPolygons} object.
}
\description{
Outlines each unique element in a \emph{RasterLayer} and returns a \emph{SpatialPolygons} object.
}
\details{
{builds a \emph{SpatialPolygons} object defined as the convex hull of a set of coordinate pairs provided through \emph{x}.}
}
\examples{
{

# read reference data
data(fieldData)

# extract centroids for each polygon
cp <- spCentroid(fieldData)

# build single polygon from centroid points
p <- simpleTrace(cp)

# compare objects
plot(p)
points(cp, col="red")

}
}
