% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecDistance.R
\name{ecDistance}
\alias{ecDistance}
\title{ecDistance}
\usage{
ecDistance(x)
}
\arguments{
\item{x}{A \emph{matrix}, \emph{data.frame} or a \emph{SpatialPoints} object.}
}
\value{
A \emph{matrix}.
}
\description{
Calculates the ecludian distance among all elements of a SpatialPoints object.
}
\details{
{compares all elements of \emph{x} and returns the minimum ecludian distance between them.}
}
\examples{
{}
}
\seealso{
\code{\link{checkOverlap}}
}
