% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTS.R
\name{extractTS}
\alias{extractTS}
\title{extractTS}
\usage{
extractTS(x, y)
}
\arguments{
\item{x}{Object of class \emph{SpatialPolygons} or \emph{SpatialPolygonsDataFrame}.}

\item{y}{A \emph{raster} object or a numeric element.}
}
\value{
A \emph{list}.
}
\description{
Summarizes multi-band \emph{raster} data within each element of a \emph{SpatialPolygons} object.
}
\details{
{For each polygon in \emph{x}, the function identifies the overlapping pixels in \emph{y} and, for each pixel, estimates the
percentage area covered by the polygon. Using this data as weights, the function calculates the weighted mean for each band in \emph{y}. 
If \emph{y} is a numeric element, the function will build a raster with resolution equal to \emph{y} over which the pixel cover will be 
estimated. The function returns a list of three \emph{data.frame} objects where each row represents a different polygon in \emph{x}:
\itemize{
\item{\emph{pixel.info} - Extracted weighted-mean time-series.}
\item{\emph{polygon.info} - Mean, min, max and standard deviation of the pixel cover; centroid coordinates.}
\item{\emph{weighted.mean} - Weighted mean raster values (if \emph{y} is a raster object).}}}
}
\examples{
{

# read raster data
r <- brick(system.file("extdata", "ndvi.tif", package="fieldRS"))

# read field data
p <- shapefile(system.file("extdata", "fields.shp", package="fieldRS"))

# derive time series
ev <- extractTS(r, p)

# see information on selected pixels
head(ev$pixel.info)

# plot profile
plot (ec$weighted.mean, type="l")

}
}
\seealso{
\code{\link{analyzeTS}}
}
