% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldRS.R
\docType{data}
\name{fieldData}
\alias{fieldData}
\title{Polygon shapefile.}
\format{
A SpatialPolygonsDataFrame
}
\usage{
data(fieldData)
}
\description{
Ground truth data on crop types collected in Uzbekistan.
}
\details{
\itemize{
  \item{crop}{Crop type.}
  \item{date}{Sampling date.}
  \item{area}{Area in m^{2}.}
}
}
\keyword{datasets}
