% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixFilter.R
\name{pixFilter}
\alias{pixFilter}
\title{pixFilter}
\usage{
pixFilter(x, y, method)
}
\arguments{
\item{x}{Object of class \emph{RasterLayer}.}

\item{y}{A \emph{numeric} element.}

\item{z}{A \emph{character} element. One of "erosion" or "dilation".}
}
\value{
A \emph{RasterLayer}.
}
\description{
Erosion and dilation filter of a raster image.
}
\examples{
{

require(raster)

# read raster data
r <- raster(system.file("extdata", "ndvi.tif", package="fieldRS"))

# filter image (3x3 erosion)
or <- pixFilter(r, 1, "erode")
plot(r)
plot(or)

#' # filter image (3x3 dilation)
or <- pixFilter(r, 1, "dilate")
plot(r)
plot(or)

}
}
