% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldRS.R
\docType{data}
\name{predictive.model1}
\alias{predictive.model1}
\title{Predictive classification model accuracy (class-wise)}
\format{A logical vector.}
\usage{
data(predictive.model1)
}
\description{
Sample-wise validation returned by classModel().
}
\keyword{datasets}
