% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldRS.R
\docType{data}
\name{predictive.model2}
\alias{predictive.model2}
\title{Predictive classification model accuracy (overall)}
\format{A data.frame}
\usage{
data(predictive.model2)
}
\description{
Class-wise F1-score returned by classModel().
}
\keyword{datasets}
