% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldRS.R
\docType{data}
\name{referenceProfiles}
\alias{referenceProfiles}
\title{Reference profiles.}
\format{
A data.frame
}
\usage{
data(referenceProfiles)
}
\description{
Reference NDVI profiles of selected classes.
}
\keyword{datasets}
