% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative.freq.R
\name{relative.freq}
\alias{relative.freq}
\title{relative.freq}
\usage{
relative.freq(x, na.rm = TRUE)
}
\arguments{
\item{x}{A vector of class \emph{numeric} or an object of class \emph{rasterLayer}.}

\item{na.rm}{Logical. Should the NA values be excluded. Default is TRUE.}
}
\value{
A \emph{numeric} element or an object of class \emph{RasterLayer}.
}
\description{
Estimate the relative frequency of non-NA pixels.
}
\details{
{Estimates the relative frequency of non-NA values.}
}
\examples{
{

x <- c(1, 1, 1, NA, NA, 1, NA)
f <- relative.freq(x, na.rm=TRUE)

}
}
\seealso{
\code{\link{mape}}
}
