% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldRS.R
\docType{data}
\name{roads}
\alias{roads}
\title{Road shapefile.}
\format{A SpatialLinesDataFrame}
\usage{
data(roads)
}
\description{
Road shapefile information extract from Open Street Map.
}
\keyword{datasets}
