% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldRS.R
\docType{data}
\name{samples1}
\alias{samples1}
\title{Pixel sample shapefile.}
\format{A SpatialLinesDataFrame}
\usage{
data(samples1)
}
\description{
Samples derived from the fieldData dataset with poly2sample().
}
\keyword{datasets}
