% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleTrace.R
\name{simpleTrace}
\alias{simpleTrace}
\title{simpleTrace}
\usage{
simpleTrace(x)
}
\arguments{
\item{x}{Object of class \emph{data.frame}, \emph{matrix}, \emph{SpatialPoints} or \emph{SpatialPointsDataFrame}.}
}
\value{
A \emph{SpatialPolygons} object.
}
\description{
Outlines a set of spatial points and returns a \emph{SpatialPolygons} object.
}
\details{
{builds a \emph{SpatialPolygons} object defined as the convex hull of a set of coordinate pairs provided through \emph{x}.}
}
\examples{
{

# read reference data
data(fieldData)

# extract centroids for each polygon
cp <- spCentroid(fieldData)

# build single polygon from centroid points
p <- simpleTrace(cp)

# compare objects
plot(p)
points(cp, col="red")

}
}
