% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothTrace.R
\name{smoothTrace}
\alias{smoothTrace}
\title{smoothTrace}
\usage{
smoothTrace(x, f)
}
\arguments{
\item{x}{Object of class \emph{RasterLayer}.}

\item{f}{Object of class \emph{numeric}.}
}
\value{
A \emph{SpatialPolygons} object.
}
\description{
Outlines each unique element in a \emph{RasterLayer} and returns a \emph{SpatialPolygons} object.
}
\details{
{Aggregates \emph{x} to a lower spatial resolution, with a factor defined by \emph{f}, an disaggregates 
the output back to the native resolution. Finally, the function crops the output using the extent of the non-NA 
valuesin \emph{x} and derives a \emph{SpatialPolygons} object.}
}
\examples{
{

require(raster) 

# read reference data
data(fieldData)

# extract centroids for each polygon
cp <- spCentroid(fieldData)

# build single polygon from centroid points (simple)
p1 <- simpleTrace(cp)

# build polygon from raster (smooth)
p2 <- smoothTrace(rasterize(p1, raster(extent(p1), res=30, crs=crs(p1))), 1)

# compare objects
plot(p1)
points(cp, col="red")
plot(p2, add=TRUE, border="blue")

}
}
