% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignClass.R
\name{assignClass}
\alias{assignClass}
\title{assignClass}
\usage{
assignClass(x, y, z)
}
\arguments{
\item{x}{Object of class \emph{data.frame}.}

\item{y}{Object of class \emph{data.frame}.}

\item{z}{A character vector.}
}
\value{
A t\emph{data.frame}.
}
\description{
Classifies samples based on reference data.
}
\details{
{For each row in \emph{x}, the function derives the coefficient of determination (\eqn{2^{x}}) between it and each 
row in \emph{y}. Then, the function will identify the value in \emph{z} that is associated with the highest \eqn{2^{x}}. The 
output if a \emph{data.frame} reporting on the \eqn{2^{x}}) and on the selected \emph{z} value.}
}
\examples{
{}
}
\seealso{
\code{\link{extractTS}} \code{\link{analyzeTS}} \code{\link{classModel}}
}
