% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignClass.R
\name{assignClass}
\alias{assignClass}
\title{assignClass}
\usage{
assignClass(x, y, z)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A \emph{matrix} or a \emph{data.frame}.}
}
\value{
A list containing containing multi-variate information on the classification of \emph{x}.
}
\description{
Classifies temporal profiles based on reference data.
}
\details{
{The function correlates \emph{x} against each row in \emph{y} and classifies
\emph{x} based on the highest r2 value. The final output is provided as a list containing:
\itemize{
 \item{\emph{class} - Class ID of \emph{x} corresponding to the rrow in \emph{y} with the highest r2 when compared with \emph{x}.}
 \item{\emph{r2} - R2 value for the best fit.}
 \item{\emph{r2diff} - Difference between the 1st and 2nd highest correlation values.}
 \item{\emph{rmse} - Root Mean Square Error (RMSE) betwen \emph{x} and the best fitting profile in \emph{y}.}}}
}
\examples{
{

r <- brick(system.file("extdata", "ndvi.tif", package="fieldRS"))

# read field data
p <- shapefile(system.file("extdata", "fields.shp", package="fieldRS"))

# derive time series
ev <- extractTS(r, p)


}
}
\seealso{
\code{\link{extractTS}} \code{\link{analyzeTS}}
}
