% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccLabel.R
\name{ccLabel}
\alias{ccLabel}
\title{ccLabel}
\usage{
ccLabel(x, method = "simple", change.threshold = NULL)
}
\arguments{
\item{x}{Object of class \emph{RasterLayer}, \emph{RasterStack} or \emph{RasterBrick}.}

\item{method}{Labeling method. Choose between 'simple' and 'change'. Default is 'simple'.}

\item{change.threshold}{Numeric element.}
}
\value{
A list.
}
\description{
Labels groups of pixels in a raster object that share similar atributes.
}
\details{
{Uses a 8-neighbor connected compoent labelling algorithm (determined by \emph{method}) to identify groups of pixels of the same
value. Each group receives a distinct numeric label. The function provides two connected component labeling alogorithms:
\itemize{
 \item{\emph{simple} - Connects neighboring pixels with the same value. Suitable for categorical data.}
 \item{\emph{space_change} - Estimates the MADE using a 3x3 moving window and connects pixels where the 
 spatial change \emph{change.threshold}.}
 #'  \item{\emph{time_change} - Estimates the MADE using all layers in a multi-band raster and connects 
 pixels where the temporal change is below \emph{change.threshold}.}}
The final output of the function is a list consisting of:
\itemize{
 \item{\emph{regions} - \emph{RasterLayer} object with region labels.}
 \item{\emph{frequency} - \emph{data.frame} object with the pixel count for each unique value in \emph{regions}.}}}
}
\seealso{
\code{\link{classModel}} \code{\link{classFilter}}
}
