% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkOverlap.R
\name{checkOverlap}
\alias{checkOverlap}
\title{checkOverlap}
\usage{
checkOverlap(x, y)
}
\arguments{
\item{x}{A spatial object.}

\item{y}{A spatial object.}
}
\value{
A two element numeric \emph{vector}.
}
\description{
Reports on how much two spatial objects overlap.
}
\details{
{Uses \link[raster]{intersect} to estimate the amount of overlap between two spatial objects (i.e. \emph{x}
and \emph{y}). The function reports on the percentage of the area in each spatial object represented by their overlap.}
}
\examples{
{}
}
\seealso{
\code{\link{derivePlots}} \code{\link{rankPlots}}
}
