% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTS.R
\name{extractTS}
\alias{extractTS}
\title{extractTS}
\usage{
extractTS(x, y)
}
\arguments{
\item{x}{Object of class \emph{SpatialPolygons} or \emph{SpatialPolygonsDataFrame}.}

\item{y}{A \emph{raster} object.}
}
\value{
A \emph{data.frame}.
}
\description{
Summarizes multi-band \emph{raster} data within each element of a \emph{SpatialPolygons} object.
}
\details{
{For each element in \emph{x}, the function will identify the overlapping pixels in \emph{y} and, for each pixel, estimates the
percentage area covered by the polygon. Using this data as weights, the function estimates a weighted mean for each band in \emph{y}. The
function returns a list containing three \emph{data.frame} objects where each row represents a different polygon in \emph{x}:
\itemize {
\item{\emph{values} - Extracted weighted-mean time-series.}
\item{\emph{coords} - Polygon center coordinates.}
\item{\emph{cover} - Mean, min, max and standard deviation of the pixel cover.}}}
}
\examples{
{}
}
\seealso{
\code{\link{analyzeTS}}
}
